+++ src/modify.c
@@ -113,7 +113,7 @@ void string_add(struct descriptor_data *d, char *str)
     }
   
   if (!(*d->str))	{
-    if (strlen(str) > d->max_str)  	{
+    if (strlen(str) > (unsigned int)d->max_str)  	{
       send_to_char("String too long - Truncated.\n\r",
 		   d->character);
       *(str + d->max_str) = '\0';
@@ -122,7 +122,7 @@ void string_add(struct descriptor_data *d, char *str)
     CREATE(*d->str, char, strlen(str) + 3);
     strcpy(*d->str, str);
   } else	{
-    if (strlen(str) + strlen(*d->str) > d->max_str)	{
+    if (strlen(str) + strlen(*d->str) > (size_t)d->max_str)	{
       send_to_char("String too long. Last line skipped.\n\r",
 		   d->character);
       terminator = 1;
@@ -377,7 +377,7 @@ void do_string(struct char_data *ch, char *arg, int UNUSED(cmd)) {
   }
   
   if (*string) {   /* there was a string in the argument array */ 
-    if (strlen(string) > length[field - 1])	{
+    if (strlen(string) > (size_t)length[field - 1])	{
       send_to_char("String too long - truncated.\n\r", ch);
       *(string + length[field - 1]) = '\0';
     }
@@ -690,7 +690,7 @@ void do_edit(struct char_data *ch, char *arg, int UNUSED(cmd)) {
   }
   
   if (*string) {   /* there was a string in the argument array */ 
-    if (strlen(string) > room_length[field - 1])	{
+    if (strlen(string) > (size_t)room_length[field - 1])	{
       send_to_char("String too long - truncated.\n\r", ch);
       *(string + length[field - 1]) = '\0';
     }
